+++ lib/AppInfo/Application.php
@@ -8,6 +8,7 @@ use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
+use OCA\UserVO\Controller\AdminController;
 
 class Application extends App implements IBootstrap {
     public function __construct() {
@@ -15,7 +16,6 @@ class Application extends App implements IBootstrap {
     }
 
     public function register(IRegistrationContext $context): void {
-        //$context->registerBackendProvider(UserVOAuth::class);
     }
 
     public function boot(IBootContext $context): void {